feat(avo-2200): move route definition
