+++ ui/.env/.env.template
@@ -4,4 +4,3 @@ GRAPHQL_URL_HETARCHIEF=
 GRAPHQL_SECRET_HETARCHIEF=
 
 NPM_PUBLISH_REGISTRY=
-ADMIN_CORE_ROUTES_PREFIX=
