+++ ui/src/modules/content-page/content-page.routes.tsx
@@ -1,24 +1,9 @@
 import React, { ReactNode } from 'react';
 import { Route } from 'react-router-dom';
-import { ROUTE_PARTS } from '~modules/shared/consts/routes';
-
 import { ContentPageOverviewPage } from './ContentPageOverviewPage';
 import ContentPageEditPage from './ContentPageEditPage';
 import ContentPageDetailPage from './ContentPageDetailPage';
-import { AdminConfigManager } from '../../react-admin';
-
-export const CONTENT_PAGE_PATH = (parts = ROUTE_PARTS) => ({
-	OVERVIEW: `/${parts.admin}/${parts.content}`,
-	CREATE: `/${parts.admin}/${parts.content}/${parts.create}`,
-	DETAIL: `/${parts.admin}/${parts.content}/:id`,
-	EDIT: `/${parts.admin}/${parts.content}/:id/${parts.edit}`,
-	PAGES: `/${parts.admin}/${parts.content}?content_type=PAGINA`,
-	NEWS: `/${parts.admin}/${parts.content}?content_type=NIEUWS_ITEM`,
-	FAQS: `/${parts.admin}/${parts.content}?content_type=FAQ_ITEM`,
-	SCREENCASTS: `/${parts.admin}/${parts.content}?content_type=SCREENCAST`,
-	PROJECTS: `/${parts.admin}/${parts.content}?content_type=PROJECT`,
-	OVERVIEWS: `/${parts.admin}/${parts.content}?content_type=OVERZICHT`,
-});
+import { AdminConfigManager, CONTENT_PAGE_PATH } from '../../react-admin';
 
 export const renderAdminContentPageRoutes = (
 	parts = AdminConfigManager.getConfig().route_parts
