+++ ui/src/react-admin/modules/shared/consts/admin.const.ts
@@ -1,11 +1,11 @@
 import { NAVIGATION_PATH } from '~modules/navigation/navigation.consts';
-import { CONTENT_PAGE_PATH } from '../../../../modules/content-page/content-page.routes';
+import { AdminConfigManager, CONTENT_PAGE_PATH } from '../../../../react-admin';
 import { COLLECTIONS_OR_BUNDLES_PATH } from './collection.const';
 import { USER_PATH } from './user.const';
 
 export const ADMIN_PATH = Object.freeze({
 	...USER_PATH,
-	...CONTENT_PAGE_PATH(),
+	...CONTENT_PAGE_PATH(AdminConfigManager.getConfig().route_parts),
 	...COLLECTIONS_OR_BUNDLES_PATH,
 	...NAVIGATION_PATH,
 });
